pyproject.toml
[project]
name = "etl-pipeline"
version = "0.1.0"
description = "End-to-end ETL + ML pipeline"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dill>=0.4.0",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "pymongo[srv]==3.12",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.16.3",
]
